requirements.txt
beautifulsoup4==4.8.0
discord.py==1.3.3
requests==2.22.0





